requirements.txt
Flask_SQLAlchemy==2.5.1
Flask==1.1.2
pytest==6.2.5
flaskr==1.0.0
PyJWT==2.3.0
